Refactoring Queries modules
