+++ src/OrchardCore.Modules/Orchard.Lucene/AdminMenu.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.Extensions.Localization;
+using Microsoft.Extensions.Localization;
 using Orchard.Environment.Navigation;
 using System;
 
@@ -12,7 +12,7 @@ namespace Orchard.Lucene
         }
 
         public IStringLocalizer T { get; set; }
-        
+
         public void BuildNavigation(string name, NavigationBuilder builder)
         {
             if (!String.Equals(name, "admin", StringComparison.OrdinalIgnoreCase))
@@ -24,9 +24,12 @@ namespace Orchard.Lucene
                 .Add(T["Design"], "10", design => design
                     .AddClass("menu-design").Id("design")
                     .Add(T["Site"], "10", import => import
-                        .Add(T["Lucene Indexes"], "7", indexes => indexes
+                        .Add(T["Lucene Indices"], "7", indexes => indexes
                             .Action("Index", "Admin", new { area = "Orchard.Lucene" })
                             .Permission(Permissions.ManageIndexes)
+                            .LocalNav())
+                        .Add(T["Lucene Queries"], "8", queries => queries
+                            .Action("Index", "Query", new { area = "Orchard.Lucene" })
                             .LocalNav())))
                 .Add(T["Design"], design => design
                     .Add(T["Settings"], settings => settings
@@ -34,6 +37,7 @@ namespace Orchard.Lucene
                             .Action("Index", "Admin", new { area = "Orchard.Settings", groupId = "search" })
                             .LocalNav()
                         )));
+
         }
     }
 }
