+++ src/OrchardCore.Modules/Orchard.Lucene/Views/Query/Index.cshtml
@@ -1,4 +1,4 @@
-@model AdminIndexViewModel
+@model QueryIndexViewModel
 @using Orchard.ContentManagement;
 @inject IContentManager ContentManager
