+++ src/OrchardCore.Modules/Orchard.Lucene/Views/_ViewImports.cshtml
@@ -1,4 +1,4 @@
-﻿@*
+@*
     For more information on enabling MVC for empty projects, visit http://go.microsoft.com/fwlink/?LinkID=397860
 
 *@
@@ -7,6 +7,7 @@
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Orchard.Contents
+@addTagHelper *, Orchard.ResourceManagement
 
 @using Orchard.Lucene.ViewModels
 @using Orchard.ContentManagement.Display
