+++ src/OrchardCore.Modules/Orchard.Queries/Module.txt
@@ -6,4 +6,3 @@ Version: 2.0.x
 OrchardVersion: 2.0.x
 Description: Provides querying capabilities.
 Category: Content
-Dependencies: Orchard.Lucene
