--- src/OrchardCore.Modules/Orchard.Queries/Startup.cs
@@ -1,24 +1,0 @@
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Modules;
-using Microsoft.AspNetCore.Routing;
-using Microsoft.Extensions.DependencyInjection;
-using Orchard.Environment.Navigation;
-using System;
-
-namespace Orchard.Queries
-{
-    /// <summary>
-    /// These services are registered on the tenant service collection
-    /// </summary>
-    public class Startup : StartupBase
-    {
-        public override void ConfigureServices(IServiceCollection services)
-        {
-            services.AddScoped<INavigationProvider, AdminMenu>();
-        }
-
-        public override void Configure(IApplicationBuilder app, IRouteBuilder routes, IServiceProvider serviceProvider)
-        {
-        }
-    }
-}
