+++ src/OrchardCore.Modules/Orchard.Queries/Views/_ViewImports.cshtml
@@ -9,5 +9,4 @@
 @addTagHelper *, Orchard.Contents
 @addTagHelper *, Orchard.ResourceManagement
 
-@using Orchard.Queries.ViewModels
 @using Orchard.ContentManagement.Display
