+++ src/OrchardCore/Orchard.Lucene.Abstractions/ILuceneQueryProvider.cs
@@ -1,0 +1,10 @@
+using Lucene.Net.Search;
+using Newtonsoft.Json.Linq;
+
+namespace Orchard.Lucene
+{
+    public interface ILuceneQueryProvider
+    {
+        Query CreateQuery(IQueryDslBuilder builder, LuceneQueryContext context, string type, JObject query);
+    }
+}
