+++ src/OrchardCore/Orchard.Lucene.Abstractions/IQueryDslBuilder.cs
@@ -1,0 +1,10 @@
+using Lucene.Net.Search;
+using Newtonsoft.Json.Linq;
+
+namespace Orchard.Lucene
+{
+    public interface IQueryDslBuilder
+    {
+        Query CreateQuery(LuceneQueryContext context, JObject query);
+    }
+}
