+++ src/OrchardCore/Orchard.Lucene.Abstractions/Orchard.Lucene.Abstractions.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard1.6</TargetFramework>
+    <RootNamespace>Orchard.Lucene</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Lucene.Net.Core" Version="4.8.0-alpha" />
+  </ItemGroup>
+</Project>
