+++ src/OrchardCore/Orchard.Lucene.Core/Orchard.Lucene.Core.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard1.6</TargetFramework>
+    <RootNamespace>Orchard.Lucene</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="1.1.0" />
+    <PackageReference Include="Newtonsoft.Json" Version="10.0.2" />
+    <PackageReference Include="Lucene.Net.QueryParser" Version="4.8.0-alpha" />
+    <ProjectReference Include="..\Orchard.Lucene.Abstractions\Orchard.Lucene.Abstractions.csproj" />
+  </ItemGroup>
+
+</Project>
